Add email service as separate file because hopefully we'll replace it someday
